remove transit layer, and focus on hide and showing polylines
