+++ client/src/components/Home.jsx
@@ -2,24 +2,22 @@ import { useState, useEffect } from "react";
 import { GoogleMap, LoadScript } from "@react-google-maps/api";
 import PoiMarkers from "./PoiMarkers";
 import { fetchGTFSData } from "./transitfunction";
-import TransitMap from "./TransitMap";
 
 const Home = () => {
-  
+  const [mapInstance, setMapInstance] = useState(null)
+  const [polylines, setPolylines] = useState([])
+  const [visibleTransit, setVisibleTransit] = useState(true)
   const [pointsOfInterest, setPointsOfInterest] = useState([]);
-  const [mapInstance, setMapInstance] = useState(null);
-  const [showTransitMap, setShowTransitMap] = useState(false)
 
   // This is for creating our stops because the transit layer doesn't display them
   const [stops, setStops] = useState([]);
   // This is for creating the shapes, connecting the stops together into routes
   const [shapes, setShapes] = useState([])
   
-
   // This will be triggered on Mapload
   const onMapLoad = (map) => {
     console.log("Map Loaded: ", map)
-    setMapInstance(map);
+    setMapInstance(map)
   };
 
   const mapStyles = [
@@ -36,63 +34,109 @@ const Home = () => {
     fetchGTFSData({setStops, setShapes});
   }, []);
 
-  const apiUrl = 
-  import.meta.env.MODE === "development"
-    ? "http://localhost:8080"
-    : import.meta.env.VITE_BACKEND_URL
-
-  //fetching landmarks data from backend
   useEffect(() => {
-    fetch(`${apiUrl}/api/landmarks`)
-      .then((resp) => {
-        if (!resp.ok) {
-          throw new Error(`HTTP error! Status: ${resp.status}`);
-        }
-        return resp.json()
-      })
-      .then((data) => {
-        console.log(data);
-        setPointsOfInterest(data);
-      })
-      .catch((err) => console.error("Failed to fetch landmarks", err));
-  }, []);
+    if (!mapInstance) {
+      console.log("Map instance is NOT ready yet...")
+      return
+    }
 
-  const handleChange = () =>{
-    setMapInstance(null)
-    setShowTransitMap(!showTransitMap)
+    const delayMapInstance = setTimeout(() => {
+      if (shapes.length > 0) {
+        console.log("drawing bus route with map instance: ")
+        drawBusRoute(mapInstance, shapes)
+      }
+    }, 1000)
+
+    return () => clearTimeout(delayMapInstance)
+
+  }, [mapInstance, shapes])
+
+
+  function drawBusRoute(map, shapes) {
+
+    if (!map) {
+      console.log("Map instance is null or undefined.")
+      return
+    }
+    
+    // Grouping the shapes by their ID's in order to draw the correct route
+    const shapesByRoute = shapes.reduce((acc, shape) => {
+      if (!acc[shape.shape_id]) {
+        acc[shape.shape_id] = []
+      }
+      
+      acc[shape.shape_id].push({ lat: shape.lat, lng: shape.lng })
+      return acc
+    }, {})
+  
+    const polylinesArray = [];
+
+    // Draws each route
+    Object.keys(shapesByRoute).forEach((routeId) => {
+      const routePath = shapesByRoute[routeId];
+  
+      const routePolyLine = new google.maps.Polyline({
+        path: routePath,
+        geodesic: true,
+        strokeColor: "red",
+        strokeOpacity: .1,
+        strokeWeight: 2,
+      });
+  
+      // Adds the polylines to the map
+      routePolyLine.setMap(map);
+      polylinesArray.push(routePolyLine)
+    
+      // Adds an info window to the displayed route when clicked
+      const infoWindow = new google.maps.InfoWindow({
+        content: `<strong>${routeId}</strong>`
+      })
+  
+      routePolyLine.addListener("click", function () {
+        infoWindow.setPosition(routePath[0]) // Opens the window at the start of the route
+        infoWindow.open(map)
+      })
+    })
+    setPolylines(polylinesArray)
   }
 
+  // Toggle the visibility of all polylines
+  const togglePolylines = () => {
+    polylines.forEach((polyline) => {
+      if (polyline.getMap()) {
+        polyline.setMap(null); // Hide the polyline
+        setVisibleTransit(false)
+      } else {
+        polyline.setMap(mapInstance); // Show the polyline
+        setVisibleTransit(true)
+      }
+    });
+  };
+  
   return (
     <>
-      <button onClick={handleChange}>Transit System</button>
+      <button onClick={togglePolylines}>
+        {visibleTransit ? "Hide Transit Map" : "Show Transit Map"}
+      </button>
       <LoadScript googleMapsApiKey={import.meta.env.VITE_GOOGLE_MAPS_API_KEY}>
-        {showTransitMap ? 
-          <TransitMap 
-            mapInstance={mapInstance}
-            setMapInstance={setMapInstance}
-            mapStyles={mapStyles}
-            shapes={shapes}
-            stops={stops}
-          /> 
-          :
-          <GoogleMap
-            // This is the map component that can be customized
-            mapContainerStyle={{
+
+        <GoogleMap
+          // This is the map component that can be customized
+          mapContainerStyle={{
             width: "70vh",
             height: "70vh",
             marginLeft: "26rem",
-            }}
-            center={{ lat: 38.0406, lng: -84.5037 }}
-            zoom={11.9}
-            mapId="90d6d90b957e9186" // This helps with styling default points of interest
-            gestureHandling={"cooperative"}
-            disableDefaultUI={false}
-            onLoad={onMapLoad}
-            options={{ styles: mapStyles }}
-          >
-            <PoiMarkers pois={pointsOfInterest}/>
-          </GoogleMap>
-        }
+          }}
+          center={{ lat: 38.0406, lng: -84.5037 }}
+          zoom={11.9}
+          mapId="90d6d90b957e9186" // This helps with styling default points of interest
+          gestureHandling={"cooperative"}
+          disableDefaultUI={false}
+          onLoad={onMapLoad}
+          options={{ styles: mapStyles }}
+        >
+          <PoiMarkers setPointsOfInterest={setPointsOfInterest} pois={pointsOfInterest}/>
+        </GoogleMap>
       </LoadScript>
     </>
   );
