+++ client/src/components/transitfunction.jsx
@@ -50,47 +50,3 @@ export const fetchGTFSData = async ({setStops, setShapes}) => {
 };
 
 
-export function drawBusRoute(map, shapes) {
-
-  if (!map) {
-    console.log("Map instance is null or undefined.")
-    return
-  }
-  
-  // Grouping the shapes by their ID's in order to draw the correct route
-  const shapesByRoute = shapes.reduce((acc, shape) => {
-    if (!acc[shape.shape_id]) {
-      acc[shape.shape_id] = []
-    }
-    
-    acc[shape.shape_id].push({ lat: shape.lat, lng: shape.lng })
-    return acc
-  }, {})
-
-  // Draws each route
-  Object.keys(shapesByRoute).forEach((routeId) => {
-    const routePath = shapesByRoute[routeId];
-
-    const routePolyLine = new google.maps.Polyline({
-      path: routePath,
-      geodesic: true,
-      strokeColor: "red",
-      strokeOpacity: .1,
-      strokeWeight: 2,
-    });
-
-    // Adds the polylines to the map
-    routePolyLine.setMap(map);
-    // });
-
-    // Adds an info window to the displayed route when clicked
-    const infoWindow = new google.maps.InfoWindow({
-      content: `<strong>${routeId}</strong>`
-    })
-
-    routePolyLine.addListener("click", function () {
-      infoWindow.setPosition(routePath[0]) // Opens the window at the start of the route
-      infoWindow.open(map)
-    })
-  })
-}
